LangChain/orchestrator-azure-search-dbindex-mongodb-memory/requirements.txt
azure-core==1.26.4
azure-identity==1.13.0
azure-search-documents==11.3.0
Flask==2.3.1
requests==2.28.2
langchain==0.0.245
openai==0.27.8
load-dotenv==0.1.0
pydantic==1.10.2
python-dotenv==1.0.0
pymongo==4.4.1
tiktoken==0.4.0
waitress==2.1.2
